backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
yfinance==0.2.28
tweepy==4.14.0
feedparser==6.0.10
pandas>=2.2.0
numpy==1.26.2
plotly==5.17.0
matplotlib==3.8.2
transformers==4.36.2
torch==2.1.2
chromadb==0.4.18
google-generativeai==0.3.2
structlog==23.2.0
